Added resoloution script
Added a script to deal with changing resoloutions, currently suports 4:3 and 16:9 with some resoloutions and will do fullscreen, started adding a dropdown box to the imgui dev menu to set it from there.

currently set by using arrow keys and called in the global controller.

There is a bug with the surface where when changing screen sizes if there is any blood on the screen it would have crashed, now it just resets the surface but i need to see how to just resize the surface