Pause menu mouse and keyboard support
Applied a similar logic from the main menu to the pause menu. the user can now use mouse and keyboard input to interact with the menu and the menu should be responsive across screen sizes and gui scalings.

There is a weird bug with clicking the main menu option in the menu, when using keyboard it will bring the user to the main menu but when using mouse it will restart the level, not sure as to why this is happening as the other menu functationaly works as expected