Fixed small pause menu bug
When you where on the quit to desktop dialog you could press escape and open another pause menu. i have added a variable to track if opeining the pause menu is blocked.
I also moved the logic for closing pause from the level controller to the pause gui menu, so you can now backout from quit to desktop to the pause menu, instead of straight back to the game, need to apply same logic to settings