Started working on method to generate collision for levels
I think I can generate collision for a level without manually placing them, I need to generate an array of size n*m where n and m are the width and height of the level tilemap respectivley, the array should contain 1's and 0's where 1 is a tile and 0 is empty space.

i should then be able to use an algorithm to fill in the 1's with as little number of rectangles as possible and usingh these rectangle spawn and scale collision object to match them accordingly.

I have started working on the first function to generate this array