Implemented auto-gen collision
Taking the rectangles generated from the collision map from the previous pushes i added a script to instantiate and scale collision objects to thier respective place for the level. This is done by the level controller and when finished will spawn the player and init the player camera. I have added a new object obj_player_spawn to sit in the place of where to the player would until this generation is complete.

I have applied this to all the test levels and it seems to work as intended